% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_fmt_naps_data.R
\name{archive_fmt_naps_data}
\alias{archive_fmt_naps_data}
\title{Archive formatted NAPS data}
\usage{
archive_fmt_naps_data(
  db,
  naps_data,
  fmt_data_tbl = "fmt_data",
  fmt_meta_tbl = "fmt_meta"
)
}
\arguments{
\item{db}{A duckdb connection.}

\item{naps_data}{Output of \code{\link[=get_naps_data]{get_naps_data()}} - a list of lists with the raw data and header for each NAPS data file.}

\item{fmt_data_tbl}{The name of the table to write the formatted data to.}

\item{fmt_meta_tbl}{The name of the table to write the formatted meta data to.}
}
\description{
Archive formatted NAPS data
}
