% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_raw_naps_data.R
\name{archive_raw_naps_data}
\alias{archive_raw_naps_data}
\title{Archive raw NAPS data}
\usage{
archive_raw_naps_data(
  db,
  naps_data,
  raw_data_tbl = "raw_data",
  raw_headers_tbl = "raw_data_headers"
)
}
\arguments{
\item{db}{A duckdb connection.}

\item{naps_data}{Output of \code{\link[=get_naps_data]{get_naps_data()}} - a list of lists with the raw data and header for each NAPS data file.}

\item{raw_data_tbl}{The name of the table to write the raw data to. "_v1" and "_v2" will be appended to diffentiate the two versions of the files.}

\item{raw_headers_tbl}{The name of the table to write the raw header data to (each NAPS files has a pre-data header with file info).}
}
\description{
Archive raw NAPS data
}
