% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_date_alignment.R
\name{check_date_alignment}
\alias{check_date_alignment}
\title{Check date alignment of NAPS data}
\usage{
check_date_alignment(
  aligned_data,
  pollutant,
  value_cols,
  name_cols,
  save_issues_to = NULL
)
}
\arguments{
\item{aligned_data}{Data frame containing NAPS data aligned with another source by naps_id and date.}

\item{pollutant}{Name of pollutant to check date alignment for.}

\item{value_cols}{Vector of two column names for the pollutant in \code{aligned_data}, 1 for each data source.}

\item{save_issues_to}{Optional path to save issue csvs to.}
}
\value{
List of logical vectors for each issue type (overall, site, annual, annual_site) - \code{TRUE} where no issues were found.
}
\description{
Check date alignment of NAPS data
}
