% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_database.R
\name{connect_to_database}
\alias{connect_to_database}
\title{Connect to (and create if needed) a local duckdb database}
\usage{
connect_to_database(
  db_path = file.path(system.file("extdata", package = "napsreview"), "naps.duckdb"),
  create_if_needed = TRUE
)
}
\arguments{
\item{db_path}{Path to the database file.
Defaults to the \code{naps.duckdb} file in the \code{extdata} directory of the package installation.}

\item{create_if_needed}{Should the database be created if it doesn't already exist?
Defaults to \code{TRUE}.}
}
\value{
A database connection object.
}
\description{
Connect to (and create if needed) a local duckdb database
}
