% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_naps_data.R
\name{format_naps_data}
\alias{format_naps_data}
\title{Standardize NAPS data formatting}
\usage{
format_naps_data(naps_data_list)
}
\arguments{
\item{naps_data_list}{A list of the raw NAPS data and headers for a single year and pollutant.}
}
\value{
A dataframe of formatted NAPS data for a single year and pollutant.
}
\description{
Converts v1 and v2 NAPS data files to a common format,
filters out invalid entries, and
combines date and local hour and converts to POSIXct (UTC time)
}
