% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_and_archive_bcgov_data.R
\name{get_and_archive_bcgov_data}
\alias{get_and_archive_bcgov_data}
\title{Get and archive bcgov data}
\usage{
get_and_archive_bcgov_data(db, date_range)
}
\arguments{
\item{db}{A duckdb connection.}

\item{date_range}{A vector of two date/datetime objects specifying the range of dates to get data for.}
}
\description{
Get and archive bcgov data
}
