% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_naps_data.R
\name{get_naps_data}
\alias{get_naps_data}
\title{Download and load NAPS data}
\usage{
get_naps_data(
  years,
  pollutants = c("PM25", "O3", "NO2"),
  raw_data_dir = tempdir(),
  check_if_raw_exists = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{years}{A vector of integers specifying the years to download data for.}

\item{pollutants}{A character vector specifying the pollutants to download data for.}

\item{raw_data_dir}{A character string specifying the local directory where the raw
data files should be saved. Defaults to a temporary directory.}

\item{check_if_raw_exists}{A logical specifying whether to check if the file already
exists locally before trying to download it. Defaults to \code{TRUE}.}

\item{quiet}{A logical specifying whether to supress progress messages.}
}
\value{
a list of lists, each with the raw data and header for that NAPS data file
}
\description{
Download and load NAPS data
}
